changed from int to Long in all services

+++ MoneyFi-backend/MoneyFi-api-gateway/src/main/java/com/moneyfi/apigateway/repository/TokenBlackListRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.time.LocalDateTime;
 
-public interface TokenBlackListRepository extends JpaRepository<BlackListedToken, Integer> {
+public interface TokenBlackListRepository extends JpaRepository<BlackListedToken, Long> {
 
     public BlackListedToken findByToken(String token);
