+++ MoneyFi-backend/MoneyFi-goal/src/main/java/com/moneyfi/goal/repository/GoalRepository.java
@@ -9,23 +9,23 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 
 
-public interface GoalRepository extends JpaRepository<GoalModel, Integer> {
+public interface GoalRepository extends JpaRepository<GoalModel, Long> {
 
-    public List<GoalModel> findByUserId(int userId);
+    public List<GoalModel> findByUserId(Long userId);
 
     @Query("select g from GoalModel g where g.userId=:userId and g.goalName=:goalName")
-    public GoalModel findByUserIdAndGoalName(int userId, String goalName);
+    public GoalModel findByUserIdAndGoalName(Long userId, String goalName);
 
     @Modifying
     @Transactional
     @Query("Delete from GoalModel g where g.userId=:userId and g.goalName=:goalName")
-    public void deleteParticularGoalByGoalName(int userId, String goalName);
+    public void deleteParticularGoalByGoalName(Long userId, String goalName);
 
     @Query("select sum(g.currentAmount) from GoalModel g where g.userId = :userId")
 //    @Query(nativeQuery = true, value = "exec getCurrentTotalGoalIncome @userId = :userId")
-    public Double getCurrentTotalGoalIncome(int userId);
+    public Double getCurrentTotalGoalIncome(Long userId);
 
     @Query("select sum(g.targetAmount) from GoalModel g where g.userId = :userId")
 //    @Query(nativeQuery = true, value = "exec getTargetTotalGoalIncome @userId = :userId")
-    public Double getTargetTotalGoalIncome(int userId);
+    public Double getTargetTotalGoalIncome(Long userId);
 }
